@@ -2,14 +2,12 @@ package com.premar.muvi.adapter;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Toast;
 
-import com.premar.muvi.Interface.ItemClickListener;
 import com.premar.muvi.R;
 import com.premar.muvi.activity.PersonDetailActivity;
 import com.premar.muvi.model.credits.Cast;
@@ -19,7 +17,7 @@ import com.squareup.picasso.Picasso;
 
 import java.util.List;
 
-import static com.premar.muvi.constants.AppConstants.IMAGE_URL_BASE_PATH;
+import static com.premar.muvi.utils.AppConstants.IMAGE_URL_BASE_PATH;
 
 public class CastAdapter extends RecyclerView.Adapter<CastViewHolder> {
     private Context context;
@@ -46,7 +44,7 @@ public class CastAdapter extends RecyclerView.Adapter<CastViewHolder> {
         holder.castCharacter.setText(castList.get(position).getCharacter());
 
         String image_url = IMAGE_URL_BASE_PATH + castList.get(position).getProfile_path();
-        Picasso.with(context)
+        Picasso.get()
                 .load(image_url)
                 .noFade()
                 .placeholder(R.drawable.ic_person_blue)
