@@ -2,17 +2,14 @@ package com.premar.muvi.adapter;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Toast;
 
-import com.premar.muvi.Interface.ItemClickListener;
 import com.premar.muvi.R;
 import com.premar.muvi.activity.MovieDetailActivity;
-import com.premar.muvi.constants.AppConstants;
 import com.premar.muvi.model.Movie;
 import com.premar.muvi.temporary_storage.MovieCache;
 import com.premar.muvi.viewholders.MovieHomeViewHolder;
@@ -20,32 +17,32 @@ import com.squareup.picasso.Picasso;
 
 import java.util.List;
 
-public class MovieHomeAdapter extends RecyclerView.Adapter<MovieHomeViewHolder> {
-    private List<Movie> movies;
-    private int rowLayout;
-    private Context context;
-    private String IMAGE_URL = AppConstants.IMAGE_URL_BASE_PATH;
-    private String BACKDROP_URL = AppConstants.BACKDROP_URL_BASE_PATH;
+import static com.premar.muvi.utils.AppConstants.BACKDROP_URL_BASE_PATH;
+import static com.premar.muvi.utils.AppConstants.IMAGE_URL_BASE_PATH;
 
-    public MovieHomeAdapter(List<Movie> movies, int rowLayout, Context context) {
+public class MovieHomeAdapter extends RecyclerView.Adapter<MovieHomeViewHolder> {
+    private Context context;
+    private List<Movie> movies;
+
+
+    public MovieHomeAdapter(Context context, List<Movie> movies) {
         this.movies = movies;
-        this.rowLayout = rowLayout;
         this.context = context;
     }
 
     @NonNull
     @Override
     public MovieHomeViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int i) {
-        View view = LayoutInflater.from(parent.getContext()).inflate(rowLayout, parent, false);
+        View view = LayoutInflater.from(parent.getContext()).inflate(R.layout.layout_movies, parent, false);
         return new MovieHomeViewHolder(view);
     }
 
     @Override
     public void onBindViewHolder(@NonNull MovieHomeViewHolder movieHomeViewHolder, int position) {
-        String image_url = IMAGE_URL + movies.get(position).getPosterPath();
-        String backdrop_url = BACKDROP_URL + movies.get(position).getBackdropPath();
+        String image_url = IMAGE_URL_BASE_PATH + movies.get(position).getPosterPath();
+        String backdrop_url = BACKDROP_URL_BASE_PATH + movies.get(position).getBackdropPath();
 
-        Picasso.with(context)
+        Picasso.get()
                 .load(image_url)
                 .placeholder(R.drawable.ic_picture)
                 .error(R.drawable.ic_picture)
@@ -56,7 +53,7 @@ public class MovieHomeAdapter extends RecyclerView.Adapter<MovieHomeViewHolder> 
             Movie selectedMovie = movies.get(position);
             Intent detailIntent = new Intent(context, MovieDetailActivity.class);
             //save temporary the movie details
-           MovieCache.movieId = movies.get(i).getId();
+            MovieCache.movieId = movies.get(i).getId();
 
             detailIntent.putExtra("movie", selectedMovie);
             detailIntent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
