@@ -2,28 +2,23 @@ package com.premar.muvi.adapter;
 
 import android.content.Context;
 import android.content.Intent;
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.Toast;
 
-import com.premar.muvi.Interface.ItemClickListener;
 import com.premar.muvi.R;
-import com.premar.muvi.activity.MovieDetailActivity;
 import com.premar.muvi.activity.TvDetailActivity;
-import com.premar.muvi.model.Movie;
 import com.premar.muvi.model.tv.Tv;
 import com.premar.muvi.temporary_storage.MovieCache;
-import com.premar.muvi.viewholders.MovieHomeViewHolder;
 import com.premar.muvi.viewholders.TvViewHolder;
 import com.squareup.picasso.Picasso;
 
 import java.util.List;
 
-import static com.premar.muvi.constants.AppConstants.BACKDROP_URL_BASE_PATH;
-import static com.premar.muvi.constants.AppConstants.IMAGE_URL_BASE_PATH;
+import static com.premar.muvi.utils.AppConstants.BACKDROP_URL_BASE_PATH;
+import static com.premar.muvi.utils.AppConstants.IMAGE_URL_BASE_PATH;
 
 public class TvAdapter extends RecyclerView.Adapter<TvViewHolder> {
     private List<Tv> shows;
@@ -48,7 +43,7 @@ public class TvAdapter extends RecyclerView.Adapter<TvViewHolder> {
         String image_url = IMAGE_URL_BASE_PATH + shows.get(position).getPoster_path();
         String backdrop_url = BACKDROP_URL_BASE_PATH + shows.get(position).getBackdrop_path();
 
-        Picasso.with(context)
+        Picasso.get()
                 .load(image_url)
                 .placeholder(R.drawable.ic_picture)
                 .error(R.drawable.ic_picture)
