@@ -1,9 +1,9 @@
 package com.premar.muvi.fragments;
 
 
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
 
 import com.premar.muvi.fragments.movie_fragment.CastFragment;
 import com.premar.muvi.fragments.movie_fragment.CommentsFragment;
