@@ -1,8 +1,8 @@
 package com.premar.muvi.fragments;
 
-import android.support.v4.app.Fragment;
-import android.support.v4.app.FragmentManager;
-import android.support.v4.app.FragmentPagerAdapter;
+import androidx.fragment.app.Fragment;
+import androidx.fragment.app.FragmentManager;
+import androidx.fragment.app.FragmentPagerAdapter;
 
 import com.premar.muvi.fragments.tv_fragments.TvCastFragment;
 import com.premar.muvi.fragments.tv_fragments.TvInfoFragment;
