@@ -2,9 +2,9 @@ package com.premar.muvi.fragments.allmovies;
 
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.GridLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.GridLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -25,7 +25,7 @@ import retrofit2.Call;
 import retrofit2.Callback;
 import retrofit2.Response;
 
-import static com.premar.muvi.constants.AppConstants.API_KEY;
+import static com.premar.muvi.utils.AppConstants.API_KEY;
 
 /**
  * A simple {@link Fragment} subclass.
@@ -86,7 +86,7 @@ public class PopularMoviesFragment extends Fragment {
                 totalPages = movies.getTotalPages();
 
                 movieList = movies.getResults();
-                adapter = new MovieHomeAdapter(movieList, R.layout.layout_movies, getContext());
+                adapter = new MovieHomeAdapter(getContext(), movieList);
                 popular_recycleview.setAdapter(adapter);
                 Log.i(TAG, "Number of trending movies received:" + movieList.size());
             }
