@@ -3,9 +3,9 @@ package com.premar.muvi.fragments.movie_fragment;
 
 import android.content.Intent;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -15,7 +15,7 @@ import android.widget.TextView;
 import com.premar.muvi.R;
 import com.premar.muvi.activity.AllCastActivity;
 import com.premar.muvi.adapter.CastAdapter;
-import com.premar.muvi.constants.AppConstants;
+import com.premar.muvi.utils.AppConstants;
 import com.premar.muvi.model.credits.Cast;
 import com.premar.muvi.model.credits.Credits;
 import com.premar.muvi.api.ApiService;
