@@ -1,12 +1,11 @@
 package com.premar.muvi.viewholders;
 
-import android.support.annotation.NonNull;
-import android.support.v7.widget.RecyclerView;
+import androidx.annotation.NonNull;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.View;
 import android.widget.LinearLayout;
 import android.widget.TextView;
 
-import com.premar.muvi.Interface.ItemClickListener;
 import com.premar.muvi.R;
 
 public class GenreViewHolder extends RecyclerView.ViewHolder  {
