@@ -1,80 +1,108 @@
 package com.premar.muvi.model;
+import androidx.databinding.BaseObservable;
+import androidx.databinding.Bindable;
+import androidx.room.ColumnInfo;
+import androidx.room.Entity;
+import androidx.room.Ignore;
+import androidx.room.PrimaryKey;
+
 import com.google.gson.annotations.SerializedName;
 
 import java.io.Serializable;
 import java.util.ArrayList;
 import java.util.List;
 
-public class Movie implements Serializable {
+@Entity(tableName = "favorite_movies")
+public class Movie extends BaseObservable implements Serializable {
+    @ColumnInfo(name = "poster_path")
     @SerializedName("poster_path")
     private String posterPath;
 
+    @ColumnInfo(name = "adult")
     @SerializedName("adult")
     private boolean adult;
 
+    @ColumnInfo(name = "overview")
     @SerializedName("overview")
     private String overview;
 
+    @ColumnInfo(name = "release_date")
     @SerializedName("release_date")
     private String releaseDate;
 
+    @Ignore
     @SerializedName("genres")
     private List<Genre> genres;
 
+    @ColumnInfo(name = "id")
+    @PrimaryKey(autoGenerate = true)
     @SerializedName("id")
     private int id;
 
+    @ColumnInfo(name = "original_title")
     @SerializedName("original_title")
     private String originalTitle;
 
+    @ColumnInfo(name = "original_language")
     @SerializedName("original_language")
     private String originalLanguage;
 
+    @ColumnInfo(name = "title")
     @SerializedName("title")
     private String title;
 
+    @ColumnInfo(name = "backdrop_path")
     @SerializedName("backdrop_path")
     private String backdropPath;
 
+    @ColumnInfo(name = "popularity")
     @SerializedName("popularity")
     private double popularity;
 
+    @ColumnInfo(name = "vote_count")
     @SerializedName("vote_count")
     private int voteCount;
 
+    @ColumnInfo(name = "video")
     @SerializedName("video")
     private boolean video;
 
+    @ColumnInfo(name = "vote_average")
     @SerializedName("vote_average")
     private double voteAverage;
 
+    @ColumnInfo(name = "budget")
     @SerializedName("budget")
     private int budget;
 
+    @ColumnInfo(name = "homepage")
     @SerializedName("homepage")
     private String homepage;
 
+    @ColumnInfo(name = "imdb_id")
     @SerializedName("imdb_id")
     private String imdbId;
 
+    @ColumnInfo(name = "revenue")
     @SerializedName("revenue")
     private int revenue;
 
+    @ColumnInfo(name = "runtime")
     @SerializedName("runtime")
     private int runtime;
 
+    @ColumnInfo(name = "status")
     @SerializedName("status")
     private String status;
 
-
+    @Ignore
     public Movie() {}
 
-    public Movie(String posterPath, boolean adult, String overview, String releaseDate, List<Genre> genres, int id, String originalTitle, String originalLanguage, String title, String backdropPath, double popularity, Integer voteCount, boolean video, double voteAverage, int budget, String homepage, String imdbId, int revenue, int runtime, String status) {
+    public Movie(String posterPath, boolean adult, String overview, String releaseDate, int id, String originalTitle, String originalLanguage, String title, String backdropPath, double popularity, Integer voteCount, boolean video, double voteAverage, int budget, String homepage, String imdbId, int revenue, int runtime, String status) {
         this.posterPath = posterPath;
         this.adult = adult;
         this.overview = overview;
         this.releaseDate = releaseDate;
-        this.genres = genres;
         this.id = id;
         this.originalTitle = originalTitle;
         this.originalLanguage = originalLanguage;
@@ -92,6 +120,7 @@ public class Movie implements Serializable {
         this.status = status;
     }
 
+    @Bindable
     public String getPosterPath() {
         return posterPath;
     }
@@ -100,6 +129,7 @@ public class Movie implements Serializable {
         this.posterPath = posterPath;
     }
 
+    @Bindable
     public boolean isAdult() {
         return adult;
     }
@@ -108,6 +138,7 @@ public class Movie implements Serializable {
         this.adult = adult;
     }
 
+    @Bindable
     public String getOverview() {
         return overview;
     }
@@ -116,6 +147,7 @@ public class Movie implements Serializable {
         this.overview = overview;
     }
 
+    @Bindable
     public String getReleaseDate() {
         return releaseDate;
     }
@@ -124,14 +156,17 @@ public class Movie implements Serializable {
         this.releaseDate = releaseDate;
     }
 
+    @Ignore
     public List<Genre> getGenres() {
         return genres;
     }
 
+    @Ignore
     public void setGenres(List<Genre> genres) {
         this.genres = genres;
     }
 
+    @Bindable
     public int getId() {
         return id;
     }
@@ -140,6 +175,7 @@ public class Movie implements Serializable {
         this.id = id;
     }
 
+    @Bindable
     public String getOriginalTitle() {
         return originalTitle;
     }
@@ -148,6 +184,7 @@ public class Movie implements Serializable {
         this.originalTitle = originalTitle;
     }
 
+    @Bindable
     public String getOriginalLanguage() {
         return originalLanguage;
     }
@@ -156,6 +193,7 @@ public class Movie implements Serializable {
         this.originalLanguage = originalLanguage;
     }
 
+    @Bindable
     public String getTitle() {
         return title;
     }
@@ -164,6 +202,7 @@ public class Movie implements Serializable {
         this.title = title;
     }
 
+    @Bindable
     public String getBackdropPath() {
         return backdropPath;
     }
@@ -172,6 +211,7 @@ public class Movie implements Serializable {
         this.backdropPath = backdropPath;
     }
 
+    @Bindable
     public double getPopularity() {
         return popularity;
     }
@@ -180,6 +220,7 @@ public class Movie implements Serializable {
         this.popularity = popularity;
     }
 
+    @Bindable
     public Integer getVoteCount() {
         return voteCount;
     }
@@ -188,6 +229,7 @@ public class Movie implements Serializable {
         this.voteCount = voteCount;
     }
 
+    @Bindable
     public boolean isVideo() {
         return video;
     }
@@ -196,6 +238,7 @@ public class Movie implements Serializable {
         this.video = video;
     }
 
+    @Bindable
     public double getVoteAverage() {
         return voteAverage;
     }
@@ -204,6 +247,7 @@ public class Movie implements Serializable {
         this.voteAverage = voteAverage;
     }
 
+    @Bindable
     public int getBudget() {
         return budget;
     }
@@ -212,6 +256,7 @@ public class Movie implements Serializable {
         this.budget = budget;
     }
 
+    @Bindable
     public String getHomepage() {
         return homepage;
     }
@@ -220,6 +265,7 @@ public class Movie implements Serializable {
         this.homepage = homepage;
     }
 
+    @Bindable
     public String getImdbId() {
         return imdbId;
     }
@@ -228,6 +274,7 @@ public class Movie implements Serializable {
         this.imdbId = imdbId;
     }
 
+    @Bindable
     public int getRevenue() {
         return revenue;
     }
@@ -236,6 +283,7 @@ public class Movie implements Serializable {
         this.revenue = revenue;
     }
 
+    @Bindable
     public int getRuntime() {
         return runtime;
     }
@@ -244,6 +292,7 @@ public class Movie implements Serializable {
         this.runtime = runtime;
     }
 
+    @Bindable
     public String getStatus() {
         return status;
     }
